feat: added a separate command for testing oss & cloud tests
