+++ cypress/e2e/oss/scrapers.test.ts
@@ -1,4 +1,4 @@
-import {Organization, Bucket} from '../../src/types'
+import {Organization, Bucket} from '../../../src/types'
 
 // a generous commitment to delivering this page in a loaded state
 const PAGE_LOAD_SLA = 10000
