+++ cypress/e2e/shared/checks.test.ts
