+++ cypress/e2e/shared/communityTemplates.test.ts
