+++ cypress/e2e/shared/notificationEndpoints.test.ts
