+++ cypress/e2e/shared/queryBuilder.test.ts
