+++ cypress/e2e/shared/variables.test.ts
